% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pff_utils.R
\name{pff_check_install_pkgs}
\alias{pff_check_install_pkgs}
\title{Checks and installs a single package}
\usage{
pff_check_install_pkgs(pkg.in, my.available.packages,
  my.library.path = .libPaths()[1],
  my.repository = "https://cloud.r-project.org")
}
\arguments{
\item{pkg.in}{Name of the package to be installed}

\item{my.available.packages}{Names of locally available (installed) packages}

\item{my.library.path}{Library to search for installed packages (default = .libPaths()[1])}

\item{my.repository}{Url of CRAN repository (default = "https://cloud.r-project.org" )}
}
\value{
A dataframe with information about the result of the installation
}
\description{
This function will check if input package in pkg.in is installed and, if not, installs it from a chosen repository
}
\examples{
pff_check_install_pkgs('dplyr')
}
