% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pff_find_install_pkgs.R
\name{pff_find_and_install_pkgs}
\alias{pff_find_and_install_pkgs}
\title{Finds and installs packages used in R related files}
\usage{
pff_find_and_install_pkgs(
  folder.in,
  do.recursive = TRUE,
  my.library.path = .libPaths()[1],
  my.repository = "https://cloud.r-project.org"
)
}
\arguments{
\item{folder.in}{Folder to search for R related files}

\item{do.recursive}{Find files in all subdirectories? (default = TRUE)}

\item{my.library.path}{Library to search for installed packages (default = .libPaths()[1])}

\item{my.repository}{Url of CRAN repository (default = "https://cloud.r-project.org" )}
}
\value{
Messages in prompt and dataframe with the results
}
\description{
This function will search for all R related files in input folder.in and find loaded packages with the
library() and require() functions.
}
\examples{

\dontrun{
my.dir <- dirname(system.file('extdata/Example_Script_1.R', package = 'PkgsFromFiles'))
df.out <- pff_find_and_install_pkgs(folder.in = my.dir)
print(df.out)
}
}
