% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pff_utils.R
\name{pff_find_packages_from_folder}
\alias{pff_find_packages_from_folder}
\title{Checks for installed packages based on local library folder}
\usage{
pff_find_packages_from_folder(lib.path = .libPaths()[1])
}
\arguments{
\item{lib.path}{Path of libraries files}
}
\value{
Char vector of installed packages
}
\description{
Checks for installed packages based on local library folder
}
\examples{
my.pkgs <- pff_find_packages_from_folder()
}
