% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pff_utils.R
\name{pff_find_pkgs_from_file}
\alias{pff_find_pkgs_from_file}
\title{Finds used packages from a single file}
\usage{
pff_find_pkgs_from_file(f.in)
}
\arguments{
\item{f.in}{File to search for pattern}
}
\value{
A character object with all packages, separated by ' ; '
}
\description{
This function will search the names of R packages used in the script
}
\examples{

my.f <- system.file('extdata/Example_Script_1.R', package = 'PkgsFromFiles')
pkg.out <- pff_find_pkgs_from_file(my.f)
}
