% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pff_plot_summary_pkgs.R
\name{pff_plot_summary_pkgs}
\alias{pff_plot_summary_pkgs}
\title{Plots a summary of used packages fround in files from a given folder}
\usage{
pff_plot_summary_pkgs(folder.in, n.most.used.pkgs = 10, do.recursive = TRUE)
}
\arguments{
\item{folder.in}{Folder to search for R related files}

\item{n.most.used.pkgs}{Number of most used packages to show in plot (default = 10)}

\item{do.recursive}{Find files in all subdirectories? (default = TRUE)}
}
\value{
A ggplot2 object that can be printed with print()
}
\description{
Plots a summary of used packages fround in files from a given folder
}
\examples{
 my.dir <- dirname(system.file('extdata', package = 'PkgsFromFiles'))
 p <- pff_plot_summary_pkgs(my.dir)
 print(p)
}
